WIP. File upload and Input Extension
File upload works, I believe, but not for larger files. Maybe just due to context of the file.

Currently nothing works, the input/output situation is broken. I made some changes while testing the file upload, to allow it to handle the return of image reponses and not only text reponses. That broken everything.

Check change history and potentially reverse some of it...